Программа исправлена в соответствии с комментариями в пул реквесте
