+++ src/ru/yandex/javacource/korolkov/schedule/history/HistoryManager.java
@@ -1,6 +1,5 @@
 package ru.yandex.javacource.korolkov.schedule.history;
 
-import ru.yandex.javacource.korolkov.schedule.manager.Node;
 import ru.yandex.javacource.korolkov.schedule.task.Task;
 
 import java.util.List;
@@ -10,5 +9,5 @@ public interface HistoryManager {
 
     void add(Task task);
 
-    void removeNode(Node node);
+    void remove(int id);
 }
